Implemented Bonus tasks
BONUS​ ​TASK​ ​1
Make sure anyone that registers their product receives a confirmation
email
BONUS​ ​TASK​ ​2
Implement a Validation Script that requires the customer to enter a
valid Product Code according to the
following condition: 9​ ​characters​ ​→​ ​ ​The​ ​first​ ​3​
​characters​ ​should​ ​be​ ​letters​ ​and​ ​the​ ​other​ ​6​ ​should
be​ ​numbers.
BONUS​ ​TASK​ ​3
Implement an Autocomplete function on the Country field in the
Registration Form. You should be able to
type in the name of the country instead of clicking a drop-down list and
as you type a matching country a
suggestion should appear and be selectable.